move input_text plugins into main source
