+++ plugin/input_text/default_input_text/Cargo.toml
@@ -9,10 +9,4 @@ license = "Apache-2.0"
 crate-type = ["cdylib"]
 
 [dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
 sudachi = { path = "../../../sudachi" }
-unicode-normalization = "0.1.19"
-
-[dev-dependencies]
-tempfile = "3"
