--- plugin/input_text/default_input_text/tests/resources/rewrite.def
@@ -1,1022 +1,0 @@
-# ignore normalize list
-#   ^{char}%n
-Ⅰ
-Ⅱ
-Ⅲ
-Ⅳ
-Ⅴ
-Ⅵ
-Ⅶ
-Ⅷ
-Ⅸ
-Ⅹ
-Ⅺ
-Ⅻ
-Ⅼ
-Ⅽ
-Ⅾ
-Ⅿ
-ⅰ
-ⅱ
-ⅲ
-ⅳ
-ⅴ
-ⅵ
-ⅶ
-ⅷ
-ⅸ
-ⅹ
-ⅺ
-ⅻ
-ⅼ
-ⅽ
-ⅾ
-ⅿ
-⺀
-⺁
-⺂
-⺃
-⺄
-⺅
-⺆
-⺇
-⺈
-⺉
-⺊
-⺋
-⺌
-⺍
-⺎
-⺏
-⺐
-⺑
-⺒
-⺓
-⺔
-⺕
-⺖
-⺗
-⺘
-⺙
-⺛
-⺜
-⺝
-⺞
-⺟
-⺠
-⺡
-⺢
-⺣
-⺤
-⺥
-⺦
-⺧
-⺨
-⺩
-⺪
-⺫
-⺬
-⺭
-⺮
-⺯
-⺰
-⺱
-⺲
-⺳
-⺴
-⺵
-⺶
-⺷
-⺸
-⺹
-⺺
-⺻
-⺼
-⺽
-⺾
-⺿
-⻀
-⻁
-⻂
-⻃
-⻄
-⻅
-⻆
-⻇
-⻈
-⻉
-⻊
-⻋
-⻌
-⻍
-⻎
-⻏
-⻐
-⻑
-⻒
-⻓
-⻔
-⻕
-⻖
-⻗
-⻘
-⻙
-⻚
-⻛
-⻜
-⻝
-⻞
-⻟
-⻠
-⻡
-⻢
-⻣
-⻤
-⻥
-⻦
-⻧
-⻨
-⻩
-⻪
-⻫
-⻬
-⻭
-⻮
-⻯
-⻰
-⻱
-⻲
-⻳
-⼀
-⼁
-⼂
-⼃
-⼄
-⼅
-⼆
-⼇
-⼈
-⼉
-⼊
-⼋
-⼌
-⼍
-⼎
-⼏
-⼐
-⼑
-⼒
-⼓
-⼔
-⼕
-⼖
-⼗
-⼘
-⼙
-⼚
-⼛
-⼜
-⼝
-⼞
-⼟
-⼠
-⼡
-⼢
-⼣
-⼤
-⼥
-⼦
-⼧
-⼨
-⼩
-⼪
-⼫
-⼬
-⼭
-⼮
-⼯
-⼰
-⼱
-⼲
-⼳
-⼴
-⼵
-⼶
-⼷
-⼸
-⼹
-⼺
-⼻
-⼼
-⼽
-⼾
-⼿
-⽀
-⽁
-⽂
-⽃
-⽄
-⽅
-⽆
-⽇
-⽈
-⽉
-⽊
-⽋
-⽌
-⽍
-⽎
-⽏
-⽐
-⽑
-⽒
-⽓
-⽔
-⽕
-⽖
-⽗
-⽘
-⽙
-⽚
-⽛
-⽜
-⽝
-⽞
-⽟
-⽠
-⽡
-⽢
-⽣
-⽤
-⽥
-⽦
-⽧
-⽨
-⽩
-⽪
-⽫
-⽬
-⽭
-⽮
-⽯
-⽰
-⽱
-⽲
-⽳
-⽴
-⽵
-⽶
-⽷
-⽸
-⽹
-⽺
-⽻
-⽼
-⽽
-⽾
-⽿
-⾀
-⾁
-⾂
-⾃
-⾄
-⾅
-⾆
-⾇
-⾈
-⾉
-⾊
-⾋
-⾌
-⾍
-⾎
-⾏
-⾐
-⾑
-⾒
-⾓
-⾔
-⾕
-⾖
-⾗
-⾘
-⾙
-⾚
-⾛
-⾜
-⾝
-⾞
-⾟
-⾠
-⾡
-⾢
-⾣
-⾤
-⾥
-⾦
-⾧
-⾨
-⾩
-⾪
-⾫
-⾬
-⾭
-⾮
-⾯
-⾰
-⾱
-⾲
-⾳
-⾴
-⾵
-⾶
-⾷
-⾸
-⾹
-⾺
-⾻
-⾼
-⾽
-⾾
-⾿
-⿀
-⿁
-⿂
-⿃
-⿄
-⿅
-⿆
-⿇
-⿈
-⿉
-⿊
-⿋
-⿌
-⿍
-⿎
-⿏
-⿐
-⿑
-⿒
-⿓
-⿔
-⿕
-豈
-更
-車
-賈
-滑
-串
-句
-龜
-龜
-契
-金
-喇
-奈
-懶
-癩
-羅
-蘿
-螺
-裸
-邏
-樂
-洛
-烙
-珞
-落
-酪
-駱
-亂
-卵
-欄
-爛
-蘭
-鸞
-嵐
-濫
-藍
-襤
-拉
-臘
-蠟
-廊
-朗
-浪
-狼
-郎
-來
-冷
-勞
-擄
-櫓
-爐
-盧
-老
-蘆
-虜
-路
-露
-魯
-鷺
-碌
-祿
-綠
-菉
-錄
-鹿
-論
-壟
-弄
-籠
-聾
-牢
-磊
-賂
-雷
-壘
-屢
-樓
-淚
-漏
-累
-縷
-陋
-勒
-肋
-凜
-凌
-稜
-綾
-菱
-陵
-讀
-拏
-樂
-諾
-丹
-寧
-怒
-率
-異
-北
-磻
-便
-復
-不
-泌
-數
-索
-參
-塞
-省
-葉
-說
-殺
-辰
-沈
-拾
-若
-掠
-略
-亮
-兩
-凉
-梁
-糧
-良
-諒
-量
-勵
-呂
-女
-廬
-旅
-濾
-礪
-閭
-驪
-麗
-黎
-力
-曆
-歷
-轢
-年
-憐
-戀
-撚
-漣
-煉
-璉
-秊
-練
-聯
-輦
-蓮
-連
-鍊
-列
-劣
-咽
-烈
-裂
-說
-廉
-念
-捻
-殮
-簾
-獵
-令
-囹
-寧
-嶺
-怜
-玲
-瑩
-羚
-聆
-鈴
-零
-靈
-領
-例
-禮
-醴
-隸
-惡
-了
-僚
-寮
-尿
-料
-樂
-燎
-療
-蓼
-遼
-龍
-暈
-阮
-劉
-杻
-柳
-流
-溜
-琉
-留
-硫
-紐
-類
-六
-戮
-陸
-倫
-崙
-淪
-輪
-律
-慄
-栗
-率
-隆
-利
-吏
-履
-易
-李
-梨
-泥
-理
-痢
-罹
-裏
-裡
-里
-離
-匿
-溺
-吝
-燐
-璘
-藺
-隣
-鱗
-麟
-林
-淋
-臨
-立
-笠
-粒
-狀
-炙
-識
-什
-茶
-刺
-切
-度
-拓
-糖
-宅
-洞
-暴
-輻
-行
-降
-見
-廓
-兀
-嗀
-﨎
-﨏
-塚
-﨑
-晴
-﨓
-﨔
-凞
-猪
-益
-礼
-神
-祥
-福
-靖
-精
-羽
-﨟
-蘒
-﨡
-諸
-﨣
-﨤
-逸
-都
-﨧
-﨨
-﨩
-飯
-飼
-館
-鶴
-郞
-隷
-侮
-僧
-免
-勉
-勤
-卑
-喝
-嘆
-器
-塀
-墨
-層
-屮
-悔
-慨
-憎
-懲
-敏
-既
-暑
-梅
-海
-渚
-漢
-煮
-爫
-琢
-碑
-社
-祉
-祈
-祐
-祖
-祝
-禍
-禎
-穀
-突
-節
-練
-縉
-繁
-署
-者
-臭
-艹
-艹
-著
-褐
-視
-謁
-謹
-賓
-贈
-辶
-逸
-難
-響
-頻
-恵
-𤋮
-舘
-並
-况
-全
-侀
-充
-冀
-勇
-勺
-喝
-啕
-喙
-嗢
-塚
-墳
-奄
-奔
-婢
-嬨
-廒
-廙
-彩
-徭
-惘
-慎
-愈
-憎
-慠
-懲
-戴
-揄
-搜
-摒
-敖
-晴
-朗
-望
-杖
-歹
-殺
-流
-滛
-滋
-漢
-瀞
-煮
-瞧
-爵
-犯
-猪
-瑱
-甆
-画
-瘝
-瘟
-益
-盛
-直
-睊
-着
-磌
-窱
-節
-类
-絛
-練
-缾
-者
-荒
-華
-蝹
-襁
-覆
-視
-調
-諸
-請
-謁
-諾
-諭
-謹
-變
-贈
-輸
-遲
-醙
-鉶
-陼
-難
-靖
-韛
-響
-頋
-頻
-鬒
-龜
-𢡊
-𢡄
-𣏕
-㮝
-䀘
-䀹
-𥉉
-𥳐
-𧻓
-齃
-龎
-゛
-゜
-
-# replace char list
-#   ^{before}\s{after}%n
-ｳﾞ	ヴ
-ｶﾞ	ガ
-ｷﾞ	ギ
-ｸﾞ	グ
-ｹﾞ	ゲ
-ｺﾞ	ゴ
-ｻﾞ	ザ
-ｼﾞ	ジ
-ｽﾞ	ズ
-ｾﾞ	ゼ
-ｿﾞ	ゾ
-ﾀﾞ	ダ
-ﾁﾞ	ヂ
-ﾂﾞ	ヅ
-ﾃﾞ	デ
-ﾄﾞ	ド
-ﾊﾞ	バ
-ﾋﾞ	ビ
-ﾌﾞ	ブ
-ﾍﾞ	ベ
-ﾎﾞ	ボ
-ﾊﾟ	パ
-ﾋﾟ	ピ
-ﾌﾟ	プ
-ﾍﾟ	ペ
-ﾎﾟ	ポ
-うﾞ	ゔ
-かﾞ	が
-きﾞ	ぎ
-くﾞ	ぐ
-けﾞ	げ
-こﾞ	ご
-さﾞ	ざ
-しﾞ	じ
-すﾞ	ず
-せﾞ	ぜ
-そﾞ	ぞ
-たﾞ	だ
-ちﾞ	ぢ
-つﾞ	づ
-てﾞ	で
-とﾞ	ど
-はﾞ	ば
-ひﾞ	び
-ふﾞ	ぶ
-へﾞ	べ
-ほﾞ	ぼ
-はﾟ	ぱ
-ひﾟ	ぴ
-ふﾟ	ぷ
-へﾟ	ぺ
-ほﾟ	ぽ
-ウﾞ	ヴ
-カﾞ	ガ
-キﾞ	ギ
-クﾞ	グ
-ケﾞ	ゲ
-コﾞ	ゴ
-サﾞ	ザ
-シﾞ	ジ
-スﾞ	ズ
-セﾞ	ゼ
-ソﾞ	ゾ
-タﾞ	ダ
-チﾞ	ヂ
-ツﾞ	ヅ
-テﾞ	デ
-トﾞ	ド
-ハﾞ	バ
-ヒﾞ	ビ
-フﾞ	ブ
-ヘﾞ	ベ
-ホﾞ	ボ
-ハﾟ	パ
-ヒﾟ	ピ
-フﾟ	プ
-ヘﾟ	ペ
-ホﾟ	ポ
-ゔ	ゔ
-が	が
-ぎ	ぎ
-ぐ	ぐ
-げ	げ
-ご	ご
-ざ	ざ
-じ	じ
-ず	ず
-ぜ	ぜ
-ぞ	ぞ
-だ	だ
-ぢ	ぢ
-づ	づ
-で	で
-ど	ど
-ば	ば
-び	び
-ぶ	ぶ
-べ	べ
-ぼ	ぼ
-ぱ	ぱ
-ぴ	ぴ
-ぷ	ぷ
-ぺ	ぺ
-ぽ	ぽ
-ヴ	ヴ
-ガ	ガ
-ギ	ギ
-グ	グ
-ゲ	ゲ
-ゴ	ゴ
-ザ	ザ
-ジ	ジ
-ズ	ズ
-ゼ	ゼ
-ゾ	ゾ
-ダ	ダ
-ヂ	ヂ
-ヅ	ヅ
-デ	デ
-ド	ド
-バ	バ
-ビ	ビ
-ブ	ブ
-ベ	ベ
-ボ	ボ
-パ	パ
-ピ	ピ
-プ	プ
-ペ	ペ
-ポ	ポ
-う゛	ゔ
-か゛	が
-き゛	ぎ
-く゛	ぐ
-け゛	げ
-こ゛	ご
-さ゛	ざ
-し゛	じ
-す゛	ず
-せ゛	ぜ
-そ゛	ぞ
-た゛	だ
-ち゛	ぢ
-つ゛	づ
-て゛	で
-と゛	ど
-は゛	ば
-ひ゛	び
-ふ゛	ぶ
-へ゛	べ
-ほ゛	ぼ
-は゜	ぱ
-ひ゜	ぴ
-ふ゜	ぷ
-へ゜	ぺ
-ほ゜	ぽ
-ウ゛	ヴ
-カ゛	ガ
-キ゛	ギ
-ク゛	グ
-ケ゛	ゲ
-コ゛	ゴ
-サ゛	ザ
-シ゛	ジ
-ス゛	ズ
-セ゛	ゼ
-ソ゛	ゾ
-タ゛	ダ
-チ゛	ヂ
-ツ゛	ヅ
-テ゛	デ
-ト゛	ド
-ハ゛	バ
-ヒ゛	ビ
-フ゛	ブ
-ヘ゛	ベ
-ホ゛	ボ
-ハ゜	パ
-ヒ゜	ピ
-フ゜	プ
-ヘ゜	ペ
-ホ゜	ポ
