--- plugin/input_text/ignore_yomigana/tests/resources/char.def
@@ -1,20 +1,0 @@
-0x0030..0x0039 NUMERIC  #0-9
-0x0041..0x005A ALPHA    #A-Z
-0x0061..0x007A ALPHA    #a-z
-0x00C0..0x00FF ALPHA  # Latin 1 #À->ÿ
-0x3041..0x309F  HIRAGANA
-0x30A1..0x30FF  KATAKANA
-0x30A1          NOOOVBOW
-0xFF66..0xFF9D  KATAKANA
-0xFF9E..0xFF9F  KATAKANA
-0x2E80..0x2EF3  KANJI # CJK Raidcals Supplement
-0x2F00..0x2FD5  KANJI
-0x3005          KANJI
-0x3007          KANJI
-0x3400..0x4DB5  KANJI # CJK Unified Ideographs Extention
-0x4E00..0x9FA5  KANJI
-0xF900..0xFA2D  KANJI
-0xFA30..0xFA6A  KANJI
-0xFF10..0xFF19 NUMERIC
-0xFF21..0xFF3A ALPHA
-0xFF41..0xFF5A ALPHA
