+++ resources/sudachi.json
@@ -2,11 +2,11 @@
     "systemDict" : "system.dic",
     "characterDefinitionFile" : "char.def",
     "inputTextPlugin" : [
-        { "class" : "../target/release/libdefault_input_text.so" },
-        { "class" : "../target/release/libprolonged_sound_mark.so",
+        { "class" : "com.worksap.nlp.sudachi.DefaultInputTextPlugin" },
+        { "class" : "com.worksap.nlp.sudachi.ProlongedSoundMarkPlugin",
           "prolongedSoundMarks": ["ー", "-", "⁓", "〜", "〰"],
           "replacementSymbol": "ー"},
-	{ "class": "../target/release/libignore_yomigana.so",
+	    { "class": "com.worksap.nlp.sudachi.IgnoreYomiganaPlugin",
           "leftBrackets": ["(", "（"],
           "rightBrackets": [")", "）"],
           "maxYomiganaLength": 4}
