--- sudachi/src/plugin/input_text.rs
@@ -1,141 +1,0 @@
-/*
- * Copyright (c) 2021 Works Applications Co., Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-use libloading::{Library, Symbol};
-use serde_json::Value;
-use std::path::Path;
-
-use crate::config::Config;
-use crate::dic::grammar::Grammar;
-use crate::input_text::Utf8InputTextBuilder;
-use crate::plugin::loader::PluginCategory;
-use crate::prelude::*;
-
-/// Trait of plugin to modify the input text before tokenization
-pub trait InputTextPlugin: Sync + Send {
-    /// Loads necessary information for the plugin
-    fn set_up(&mut self, settings: &Value, config: &Config, grammar: &Grammar)
-        -> SudachiResult<()>;
-
-    /// Rewrites input text
-    ///
-    /// builder::replace will be used inside
-    fn rewrite(&self, builder: &mut Utf8InputTextBuilder);
-}
-
-impl PluginCategory for dyn InputTextPlugin {
-    type BoxType = Box<dyn InputTextPlugin + Sync>;
-    type InitFnType = unsafe extern "C" fn() -> SudachiResult<Self::BoxType>;
-    fn configurations(cfg: &Config) -> &[Value] {
-        &cfg.input_text_plugins
-    }
-
-    fn packaged_impl(name: &str) -> Option<Self::BoxType> {
-        match name {
-            "com.worksap.sudachi.Test" => todo!(),
-            _ => None,
-        }
-    }
-
-    fn do_setup(
-        ptr: &mut Self::BoxType,
-        settings: &Value,
-        config: &Config,
-        grammar: &Grammar,
-    ) -> SudachiResult<()> {
-        ptr.set_up(settings, config, grammar)
-    }
-}
-
-/// Declare a plugin type and its constructor.
-///
-/// # Notes
-/// This works by automatically generating an `extern "C"` function with a
-/// pre-defined signature and symbol name. Therefore you will only be able to
-/// declare one plugin per library.
-#[macro_export]
-macro_rules! declare_input_text_plugin {
-    ($plugin_type:ty, $constructor:path) => {
-        #[no_mangle]
-        pub fn load_plugin() -> *mut (dyn InputTextPlugin + Sync) {
-            // make sure the constructor is the correct type.
-            let constructor: fn() -> $plugin_type = $constructor;
-
-            let object = constructor();
-            let boxed: Box<dyn InputTextPlugin + Sync> = Box::new(object);
-            Box::into_raw(boxed)
-        }
-    };
-}
-
-/// Plugin manager to handle multiple plugins
-#[derive(Default)]
-pub struct InputTextPluginManager {
-    plugins: Vec<Box<dyn InputTextPlugin + Sync>>,
-    libraries: Vec<Library>,
-}
-
-impl InputTextPluginManager {
-    pub fn load(
-        &mut self,
-        path: &Path,
-        settings: &Value,
-        config: &Config,
-        grammar: &Grammar,
-    ) -> SudachiResult<()> {
-        type PluginCreate = unsafe fn() -> *mut (dyn InputTextPlugin + Sync);
-
-        let lib = unsafe { Library::new(path) }?;
-        let load_plugin: Symbol<PluginCreate> = unsafe { lib.get(b"load_plugin") }?;
-        let mut plugin = unsafe { Box::from_raw(load_plugin()) };
-        plugin.set_up(settings, config, grammar)?;
-
-        self.plugins.push(plugin);
-        self.libraries.push(lib);
-        Ok(())
-    }
-
-    pub fn plugins(&self) -> &[Box<dyn InputTextPlugin + Sync>] {
-        &self.plugins
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.plugins.is_empty()
-    }
-}
-
-impl Drop for InputTextPluginManager {
-    fn drop(&mut self) {
-        // Plugin drop must be called before Library drop.
-        self.plugins.clear();
-        self.libraries.clear();
-    }
-}
-
-/// Load plugins based on config data
-pub fn get_input_text_plugins(
-    config: &Config,
-    grammar: &Grammar,
-) -> SudachiResult<InputTextPluginManager> {
-    let mut manager = InputTextPluginManager::default();
-
-    for plugin in &config.input_text_plugins {
-        let lib = super::get_plugin_path(plugin, config)?;
-        manager.load(lib.as_path(), plugin, config, grammar)?;
-    }
-
-    Ok(manager)
-}
