+++ sudachi/src/plugin/loader.rs
@@ -14,29 +14,16 @@
  *  limitations under the License.
  */
 
-/*
- *  Copyright (c) 2021 Works Applications Co., Ltd.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
+use libloading::{Library, Symbol};
+use serde_json::Value;
 
 use crate::config::{Config, ConfigError};
 use crate::dic::grammar::Grammar;
 use crate::error::{SudachiError, SudachiResult};
-use libloading::{Library, Symbol};
-use serde_json::Value;
-use std::path::Path;
+use crate::plugin::PluginError;
 
+/// Holds loaded plugins, whether they are bundled
+/// or loaded from DSOs
 pub struct PluginContainer<T: PluginCategory + ?Sized> {
     libraries: Vec<Library>,
     plugins: Vec<<T as PluginCategory>::BoxType>,
@@ -92,32 +79,60 @@ impl<'a, T: PluginCategory + ?Sized> PluginLoader<'a, T> {
     }
 
     fn load_plugin(&mut self, name: &str, plugin_cfg: &Value) -> SudachiResult<()> {
-        let mut plugin = match <T as PluginCategory>::packaged_impl(name) {
-            Some(p) => p,
-            None => {
-                let full_name = self.resolve_plugin_name(name);
-                self.load_plugin_from_lib(&full_name)?
-            }
-        };
+        let mut plugin =
+            // Try to load bundled plugin first, if its name looks like it
+            if let Some(stripped_name) = name.strip_prefix("com.worksap.nlp.sudachi.") {
+                if let Some(p) = <T as PluginCategory>::bundled_impl(stripped_name) {
+                    p
+                } else {
+                    return Err(SudachiError::ConfigError(ConfigError::InvalidFormat(
+                        format!("Failed to lookup bundled plugin: {}", name)
+                    )))
+                }
+            // Otherwise treat name as DSO
+            } else {
+                let candidates = self.resolve_dso_names(name);
+                self.load_plugin_from_dso(&candidates)?
+            };
 
         <T as PluginCategory>::do_setup(&mut plugin, plugin_cfg, &self.cfg, &self.grammar)?;
+        self.plugins.push(plugin);
         Ok(())
     }
 
-    fn resolve_plugin_name(&self, name: &str) -> String {
-        let resolved = self.cfg.resolve_path(name.to_owned());
+    fn resolve_dso_names(&self, name: &str) -> Vec<String> {
+        let resolved = self.cfg.resolve_plugin_paths(name.to_owned());
         resolved
     }
 
-    fn load_plugin_from_lib(
+    fn try_load_library_from(candidates: &[String]) -> SudachiResult<Library> {
+        if candidates.is_empty() {
+            return Err(SudachiError::PluginError(PluginError::InvalidDataFormat(
+                "No candidates to load library".to_owned()
+            )))
+        }
+
+        let mut last_error = libloading::Error::IncompatibleSize;
+        for p in candidates.iter() {
+            match unsafe { Library::new(p.as_str()) } {
+                Ok(lib) => return Ok(lib),
+                Err(e) => last_error = e
+            }
+        };
+        Err(SudachiError::PluginError(PluginError::Libloading {
+            source: last_error,
+            message: format!("failed to load library from: {:?}", candidates)
+        }))
+    }
+
+    fn load_plugin_from_dso(
         &mut self,
-        name: &str,
+        candidates: &[String]
     ) -> SudachiResult<<T as PluginCategory>::BoxType> {
-        let lib = unsafe { Library::new(name)? };
+        let lib = Self::try_load_library_from(candidates)?;
         let load_fn: Symbol<fn() -> SudachiResult<<T as PluginCategory>::BoxType>> =
             unsafe { lib.get(b"load_plugin")? };
-        let mut plugin = load_fn();
-
+        let plugin = load_fn();
         self.libraries.push(lib);
         plugin
     }
@@ -140,13 +155,29 @@ fn extract_plugin_class(val: &Value) -> SudachiResult<&str> {
     }
 }
 
+/// A category of Plugins
 pub trait PluginCategory {
+    /// Boxed type of the plugin. Should be Box<dyn XXXX>.
     type BoxType;
+
+    /// Type of the initialization function.
+    /// It must take 0 arguments and return `SudachiResult<Self::BoxType>`.
     type InitFnType;
+
+    /// Extract plugin configurations from the config
     fn configurations(cfg: &Config) -> &[Value];
-    fn packaged_impl(name: &str) -> Option<Self::BoxType> {
-        None
-    }
+
+    /// Create bundled plugin for plugin name
+    /// Instead of full name like com.worksap.nlp.sudachi.ProlongedSoundMarkPlugin
+    /// should handle only the short one: ProlongedSoundMarkPlugin
+    ///
+    /// com.worksap.nlp.sudachi. (last dot included) will be stripped automatically
+    /// by the loader code
+    fn bundled_impl(name: &str) -> Option<Self::BoxType>;
+
+    /// Perform initial setup.
+    /// We can't call set_up of the plugin directly in the default implementation
+    /// of this method because we do not know the specific type yet
     fn do_setup(
         ptr: &mut Self::BoxType,
         settings: &Value,
@@ -155,6 +186,9 @@ pub trait PluginCategory {
     ) -> SudachiResult<()>;
 }
 
+/// Helper function to load the plugins of a single category
+/// Should be called with turbofish syntax and trait object type:
+/// `let plugins = load_plugins_of::<dyn InputText>(...)`.
 pub fn load_plugins_of<T: PluginCategory + ?Sized>(
     cfg: &Config,
     grammar: &Grammar,
